+++ src/app/features/products/products.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { ProductComponent } from './components/product/product.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { SharedModule } from 'src/app/shared/shared.module';
+import { ProductViewComponent } from './components/product-view/product-view.component';
 
 
 const EXPORTED_COMPONENTS = [
@@ -12,6 +13,7 @@ const EXPORTED_COMPONENTS = [
   declarations: [
     ...EXPORTED_COMPONENTS,
     ProductComponent,
+    ProductViewComponent,
     
   ],
   imports: [