+++ src/app/shp-routing.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ProductListComponent } from './features/products/components/product-list/product-list.component';
+import { ProductViewComponent } from './features/products/components/product-view/product-view.component';
 
 const routes: Routes = [
   {
     path: 'products-list',
     data: { title: 'Products list' },
     component: ProductListComponent,
+  },
+  {
+    path: 'product/:productID',
+    data: { title: 'Product Item' },
+    component: ProductViewComponent,
   },
   {
     path: '**',