+++ changelog.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 
 ### Added
 - Product List routing
+- Product View component
+- Product Service - getProduct
 
 ## [0.4.0] - 2021-10-03
 ### Removed